For some time at least Ruby has pretenured Class objects, setting their age bits to the oldest immediately on allocation.

On Ruby 3.2 and earlier, objects which were referenced by an old object immediately were themselved promoted to old as well. This had the effect of implicitly always aging callcache IMEMOs as well, since they are (always?) referenced by classes.

This commit adds the same pretenuring to imemo_callcache via RVALUE_AGE_SET_CANDIDATE which is currently performed for CLASS/MODULE/ICLASS.

At GitHub we've seen significantly more time spent in GC right after boot on Ruby versions including [[Feature #19678] Don't immediately promote children of old objects](https://github.com/ruby/ruby/pull/7821). I don't know how much this PR will help but callcache objects are well represented in the long lived objects of our heaps (we end up with about 60k almost immediately), so if they do not age immediately we will need to pay for them on each GC until they become old.

---

I wrote an synthetic benchmark, where I allocate a bunch of callcaches via method calls and them measure a series of manual minor GCs.

<details>
<summary>benchmark script</summary>

```
NUM_METHODS = 1000
NUM_CLASSES = 600

class Foo
  NUM_METHODS.times do |i|
    class_eval "def foo#{i}; end"
  end

  class_eval <<~RUBY
    def call
      #{NUM_METHODS.times.map{|i| "foo#{i}" }.join(";")}
    end
  RUBY
end

classes = NUM_CLASSES.times.map do
  Class.new(Foo)
end

def benchmark
  before = Process.clock_gettime(Process::CLOCK_MONOTONIC)
  yield
  after = Process.clock_gettime(Process::CLOCK_MONOTONIC)
  puts "%.2fms" % ((after - before) * 1000.0)
end

4.times { GC.start }

GC.disable

classes.each { _1.new.call }

benchmark { GC.start(full_mark: false) }
benchmark { GC.start(full_mark: false) }
benchmark { GC.start(full_mark: false) }
benchmark { GC.start(full_mark: false) }
benchmark { GC.start(full_mark: false) }
benchmark { GC.start(full_mark: false) }
```

</details>

**Ruby 3.2**

```
$ ruby --disable-gems test_callcache.rb
30.02ms
27.07ms
0.08ms
0.05ms
0.05ms
0.05ms
```

**Ruby master**

```
$ ruby --disable-gems test_callcache.rb
21.23ms
17.22ms
18.31ms
16.97ms
0.09ms
0.06ms
```

**This branch**
```
$ ruby --disable-gems test_callcache.rb
21.90ms
16.88ms
0.09ms
0.05ms
0.05ms
0.05ms
```


---

I could see pulling the aging up a level so that we don't have this conditional in a hot path, but maybe there is a need or future intention in having it here.

cc @peterzhu2118 @eightbitraptor @byroot @tenderlove