@byroot 2023-12-05
So I was actually advocating for removing this pre-aging, as I suspect it causes a bunch of ephemeral classes to make it to the old generation.

> At GitHub we've seen significantly more time spent in GC right after boot on Ruby versions including [[Feature #19678] Don't immediately promote children of old objects](https://github.com/ruby/ruby/pull/7821).

Are you not using `Process.warmup` or https://github.com/ko1/nakayoshi_fork ? That's the kind of problems it's meant to solve.