@jhawthorn 2023-12-05
> Are you not using `Process.warmup` or https://github.com/ko1/nakayoshi_fork ? That's the kind of problems it's meant to solve.

callcaches tend to be allocated after boot as there's no good way (yet?) to warm them up. So those techniques do not help (as in the benchmark provided)