@byroot 2023-12-05
> Many callcaches tend to be allocated after boot as there's no good way (yet?) to warm them up.

Ah my bad, I'm much less familiar with this part of the VM.

Any reason why we couldn't pre-allocate them? I suppose that it would waste memory if these callsites end up never executed?

But yeah, your explanation make sense, I suppose if anything it makes more sense to age these than even T_CLASS.