@jhawthorn 2023-12-05
I'd like to explore pre-allocating callcaches, but yeah at least the naive way would be to allocate `ObjectSpace.each_object(Class).sum{ _1.instance_methods.count }` callcaches which at least for us would about 10x the size of our heap 😅 (though it's possible that could be pruned down a while bunch and also I haven't tested it).

Yeah, that was my thinking too, these could make even more sense than classes. If we decide to not pretenure Classes (which would be nice to do at least for singleton classes, for example), we could change this to instead copy the age of the class to the imemo on creation.