@jhawthorn 2023-12-09
Looking through a production heap dump I found more cases than expected (still by far the minority minority) of callcaches which weren't being referenced (anymore?) by an oldgen Class, so I'll do some more investigation to better understand before merging.