@jhawthorn 2024-07-11
Just to update this with some things I missed on a first pass. There are two types of callcaches w/r/t their lifetime: "regular" callcaches, which are cached on a class forever, and "orphan" callcaches, which are only used as inline caches (by `super` and refinements I believe) and are more likely to be replaced.

What I wrote about their lifetime matching that of the class obviously only applies to the "regular" type, so those are the only ones, if any, we should be pretenuring. I also agree it would be nice to not pretenure classes as they to may be short-lived.

Maybe the ideal version of this would be to, only when the callcache is attached to a class, copy the class age bits onto the callcache. That should avoid all downsides.

I'll close this for now.